// kservices_ts_api/src/api/domain/entities/sme.entity.ts
export type fetchRequest = {
  id: number;
};
// ...
export type updateComplaintRequest = {
  smeId : number;
  category: number;
  complaintDetail: string;
  emailId: string;
  status: number;
  subject: string;
  id: number;
  insertDateTime:string;
};

export type addComplaintRequest = {
  smeId : number;
  category: number;
  complaintDetail: string;
  emailId: string;
  status: number;
  subject: string;
  insertDateTime:string;
};